First iteration in MSI package generation with service management
