+++ Cargo.toml
@@ -23,6 +23,7 @@ tokio-util = { version = "0.7.1", features = ["codec"] }
 serde_json = "1.0.79"
 time = "0.3.9"
 ctrlc = { version = "3.2", features = ["termination"] }
+windows-service = "0.5.0"
 
 [dev-dependencies]
 tokio-test = "*"
